+++ duplocloud/schema_k8s_cron_job_spec_v1beta1.go
@@ -1,8 +1,10 @@
 package duplocloud
 
 import (
+	"fmt"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
+	"github.com/robfig/cron"
 )
 
 func cronJobSpecFieldsV1Beta1() map[string]*schema.Schema {
@@ -64,5 +66,21 @@ func cronJobSpecFieldsV1Beta1() map[string]*schema.Schema {
 			Default:     false,
 			Description: "This flag tells the controller to suspend subsequent executions, it does not apply to already started executions. Defaults to false.",
 		},
+	}
+}
+
+// nolint
+func validateCronExpression() schema.SchemaValidateFunc {
+	return func(i interface{}, k string) (s []string, es []error) {
+		v, ok := i.(string)
+		if !ok {
+			es = append(es, fmt.Errorf("expected type of '%s' to be string", k))
+			return
+		}
+		_, err := cron.ParseStandard(v)
+		if err != nil {
+			es = append(es, fmt.Errorf("'%s' should be an valid Cron expression", k))
+		}
+		return
 	}
 }
